==> 3_modeloNaivesBayes.Rmd <==
---
title: "modeloNaivesBayes"
author: "González, Jorge / Shifres, Juan"
date: "2023-09-19"
output: html_document
---
### 1) Cargar Librerias y set de datos
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)


library(tidyverse)
library(e1071)
library(caret)
library(caTools)



```

```{r}
datos <- read.csv("datos/df_filtrado.csv")
```



### 2) Preparar set de datos
```{r}


datos <- datos |> 
  select(FALLECIDO, EDAD_DIAGNOSTICO, SEPI_APERTURA, FECHA_INTERNACION, CUIDADO_INTENSIVO, CLASIFICACION_ALGORITMO) |> 
  filter(EDAD_DIAGNOSTICO>0) |> 
  mutate(grupo_etario = case_when(EDAD_DIAGNOSTICO <= 20  ~ "grupo1_Menor_20",
                                EDAD_DIAGNOSTICO >= 20 & EDAD_DIAGNOSTICO < 40  ~ "grupo2_20_40",
                                EDAD_DIAGNOSTICO >= 40 & EDAD_DIAGNOSTICO < 50  ~ "grupo3_40_50",
                                EDAD_DIAGNOSTICO >= 50 & EDAD_DIAGNOSTICO < 60  ~ "grupo4_50_60",
                                EDAD_DIAGNOSTICO >= 60 & EDAD_DIAGNOSTICO < 70  ~ "grupo5_60_70",
                                EDAD_DIAGNOSTICO >= 70 & EDAD_DIAGNOSTICO < 80  ~ "grupo6_70_80",
                                EDAD_DIAGNOSTICO >= 80 ~ "grupo7_mas_80"))
datos$grupo_etario <- as.factor(datos$grupo_etario)
datos$FALLECIDO <- as.factor(datos$FALLECIDO)

```


```{r}


datos <- datos |> 
  mutate(INTERNACION = ifelse(!is.na(FECHA_INTERNACION), "SI", "NO"))
```


```{r}
listCovid=c('Co-detección de Influenza A y B', 'Co-detección de Influenza y OVR', 'Co-detección de Influenza y SARS COV 2', 'Co-detección de Influenza y VSR', 'Co-detección de SARS COV 2 y OVR', 'Co-detección de SARS COV 2 y VSR', 'Co-detección de VSR y OVR', 'Influenza A - H1N1', 'Influenza A - H3N2', 'Influenza A - sin subtipificar', 'Influenza B - linaje Victoria', 'Influenza B - sin linaje', 'SARS-COV-2 por métodos moleculares', 'SARS-COV-2 por test de antígeno', 'Virus sincicial respiratorio (VSR)')

listSinCovid=c('Estudiado sin resultados positivos','SARS COV-2 Negativo por test de antígeno', 'Caso invalidado por epidemiología')
  
listOtrasInfecciones=c('Otro virus respiratorio', 'Virus sincicial respiratorio (VSR)')

listSinClasificar=c('Sin clasificar')

datos <- datos |> 
  mutate(CLASIFICACION = case_when(CLASIFICACION_ALGORITMO %in% listCovid ~ "Covid",
                                   CLASIFICACION_ALGORITMO %in% listSinCovid ~ "SinCovid",
                                   CLASIFICACION_ALGORITMO %in% listOtrasInfecciones ~ "OtrasInfecciones",
                                   CLASIFICACION_ALGORITMO %in% listSinClasificar ~ "SinClasificar"))

datosFinal<-datos |> 
  select(FALLECIDO, grupo_etario, INTERNACION, CUIDADO_INTENSIVO, CLASIFICACION)


```



```{r}

write_csv(datosFinal, file = "datos/datosFinal.csv")

# compruebo si hay asignación de clases
str(datosFinal)

# compruebo si hay nulos o vacíos
any(is.na(datosFinal))

```

```{r}
# Número de datos ausentes por variable

map_dbl(datosFinal, .f = function(x){sum(is.na(x))})

```

### 3) Crear set de entrenamiento y de test
```{r}


## *****************************************SPLIT del dataframe
# particiono el dataframe en 2 partes


sample <- sample.split(datosFinal$FALLECIDO, SplitRatio = .70)
#sample = muestra o muestras
train.data <- subset(datosFinal, sample==TRUE) 
test.data <- subset(datosFinal, sample==FALSE)

#se verifica que la distribución de la variable FALLECIDO es similar en el conjunto de entrenamiento y en el de test

prop.table(table(datosFinal$FALLECIDO))

prop.table(table(train.data$FALLECIDO))

prop.table(table(test.data$FALLECIDO))

```

### 4) Crear modelo de Naive Bayes
```{r}

set.seed(2023)
modelo_nB <- naiveBayes(FALLECIDO~.,train.data)
modelo_nB

#El modelo informa las probabilidades de FALLECIDO con las variables de entrada, información de las estadísticas a priori. Para probarla se utilza los datos de test con predict():

```

```{r}
#Predecir los resultados de la base test

clase <- predict(modelo_nB,test.data,type="class")

table(test.data$FALLECIDO, clase)

```

### 5) Probar modelo

```{r}
# matriz de confusión

confusionMatrix(table(test.data$FALLECIDO,clase,dnn = c("Actual", "Predicción")))

```



###### **Una vez aplicado el modelo a nuestro conjunto de entrenamiento podemos apreciar como la edad de la persona influye notoriamente en la probabilidad de muerte por covid, así como el hecho de que hayan estado internados y/o en cuidados intensivos.**

###### **Por último y viendo la matriz de confusión resultante al aplicar el modelo al resto de los datos podemos ver que nos da una eficacia del 97%, y un intervalo de confianza entre 96% al 97%.**

___

<em>Como conclusión del trabajo podemos comentar que nos ha servido fundamentalmente para entender el ciclo de vida de un proyecto de datos con machine learning, así como el flujo de trabajo necesario para llevarlo a cabo. Entendemos nuestra entrega como una primera iteración, en la cual se aprecian las dificultades que nos encontramos al avanzar con el trabajo a la vez que fuimos aprendiendo cómo debería desarrollarse.</em>
